Inline RecursiveImageWidget into notebook; bump wigglystuff to 0.5.30
Move the rule-authoring widget out of recursive_image_widget/ and into
recursive-images.py as a self-contained appendix cell, and delete the folder.
