--- recursive_image_widget/recursive_image_widget.css
@@ -1,133 +1,0 @@
-.rimg-container {
-  color-scheme: light dark;
-  --rimg-bg: #ffffff;
-  --rimg-text: #111827;
-  --rimg-border: #d1d5db;
-  --rimg-toolbar-bg: #f9fafb;
-  --rimg-btn-hover: #f3f4f6;
-  --rimg-btn-active: #e5e7eb;
-  font-family: ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont,
-    "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
-  box-sizing: border-box;
-  border: 1px solid var(--rimg-border);
-  border-radius: 0.5rem;
-  overflow: hidden;
-  background-color: var(--rimg-bg);
-  color: var(--rimg-text);
-  display: flex;
-  flex-direction: column;
-}
-
-:where(.dark, .dark-theme, [data-theme="dark"]) .rimg-container {
-  --rimg-bg: #1f2937;
-  --rimg-text: #f9fafb;
-  --rimg-border: #4b5563;
-  --rimg-toolbar-bg: #111827;
-  --rimg-btn-hover: #4b5563;
-  --rimg-btn-active: #6b7280;
-}
-
-.rimg-toolbar {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 0.375rem;
-  padding: 0.375rem 0.5rem;
-  background-color: var(--rimg-toolbar-bg);
-  border-bottom: 1px solid var(--rimg-border);
-  flex-shrink: 0;
-}
-
-.rimg-btn {
-  display: inline-flex;
-  align-items: center;
-  height: 2.25rem;
-  padding: 0 0.875rem;
-  border: 1px solid var(--rimg-border);
-  border-radius: 0.375rem;
-  background-color: var(--rimg-bg);
-  color: var(--rimg-text);
-  cursor: pointer;
-  font-size: 14px;
-  white-space: nowrap;
-  transition: background-color 0.15s ease, border-color 0.15s ease;
-}
-
-.rimg-btn:hover {
-  background-color: var(--rimg-btn-hover);
-  border-color: var(--rimg-border);
-}
-
-.rimg-btn:active {
-  background-color: var(--rimg-btn-active);
-}
-
-.rimg-stage {
-  position: relative;
-  width: 100%;
-  aspect-ratio: 1 / 1;
-  background: #ffffff
-    repeating-conic-gradient(#eef2f7 0% 25%, #ffffff 0% 50%) 0 / 28px 28px;
-  overflow: hidden;
-  touch-action: none;
-}
-
-.rimg-el {
-  position: absolute;
-  transform-origin: center center;
-  cursor: grab;
-}
-
-.rimg-img {
-  width: 100%;
-  height: 100%;
-  display: block;
-  pointer-events: none;
-}
-
-.rimg-parent {
-  outline: 2px solid #334155;
-}
-.rimg-parent .rimg-img {
-  opacity: 0.9;
-}
-
-.rimg-child {
-  outline: 1px solid #60a5fa;
-}
-.rimg-child .rimg-img {
-  opacity: 0.55;
-}
-
-.rimg-el.selected {
-  outline-width: 2px;
-  outline-color: #2563eb;
-  z-index: 5;
-}
-.rimg-el.selected .rimg-img {
-  opacity: 0.85;
-}
-
-.rimg-handle {
-  position: absolute;
-  width: 14px;
-  height: 14px;
-  box-sizing: border-box;
-  background: #2563eb;
-  border: 2px solid #ffffff;
-  border-radius: 50%;
-}
-
-.rimg-scale {
-  right: -7px;
-  bottom: -7px;
-  cursor: nwse-resize;
-}
-
-.rimg-rotate {
-  left: 50%;
-  top: -26px;
-  margin-left: -7px;
-  background: #16a34a;
-  cursor: grab;
-}
